Finance module created. More modularization
