+++ logs/app_info.log
@@ -18,3 +18,7 @@
 2024-08-20 00:07:26 - INFO - Mission details saved in mission_logs/mission_log_2024-08-20_00-07-11.txt - [mission_logs.py:117]
 2024-08-20 00:07:33 - INFO - File 'mission_logs/mission_log_2024-08-20_00-07-11.txt' encrypted successfully. - [encryption.py:37]
 2024-08-20 00:07:33 - INFO - File 'mission_logs/mission_log_2024-08-20_00-07-11.txt' encrypted successfully. - [encryption.py:37]
+2024-08-20 15:13:49 - ERROR - Error decrypting file 'mission_logs\mission_log_2024-08-20_00-40-11.txt':  - [encryption.py:60]
+2024-08-20 15:13:49 - ERROR - Error decrypting file 'mission_logs\mission_log_2024-08-20_00-40-11.txt':  - [encryption.py:60]
+2024-08-20 15:13:49 - ERROR - Error listing mission logs:  - [mission_logs.py:128]
+2024-08-20 15:13:49 - ERROR - Error listing mission logs:  - [mission_logs.py:128]
