+++ modules/settings_modules/settings_menu.py
@@ -1,6 +1,6 @@
 import questionary
 
-from ..ui import display_welcome_message, display_panel, display_exit_message
+from ..ui import display_welcome_message, display_panel, display_exit_message, clear_screen
 from modules.folder_init import logging
 from .editor_settings import editor_main
